+++ crates/homeboy-core/src/config/record.rs
@@ -7,14 +7,13 @@ use super::{SetName, SlugIdentifiable};
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Record<T> {
     pub id: String,
-    #[serde(flatten)]
-    pub value: T,
+    pub config: T,
 }
 
 impl<T: SlugIdentifiable + SetName> Record<T> {
-    pub fn new(name: String, mut value: T) -> Result<Self> {
-        value.set_name(name);
-        let id = value.slug_id()?;
-        Ok(Self { id, value })
+    pub fn new(name: String, mut config: T) -> Result<Self> {
+        config.set_name(name);
+        let id = config.slug_id()?;
+        Ok(Self { id, config })
     }
 }